updated 4-add.c file

+++ 0x0A-argc_argv/4-add.c
@@ -12,7 +12,7 @@
 int main(int argc, char *argv[])
 {
 	int i, flag;
-	unsigned long int j;
+	unsigned int j;
 	int sum = 0;
 
 	if (argc == 1)
